requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
docker==7.1.0
faiss-cpu==1.10.0
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.47
langchain-openai==0.3.9
langchain-text-splitters==0.3.7
langsmith==0.3.18
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.2.0
mypy-extensions==1.0.0
narwhals==1.31.0
networkx==3.4.2
numpy==2.2.4
openai==1.68.0
orjson==3.10.15
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.1.0
propcache==0.3.0
protobuf==5.29.4
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyMuPDF==1.25.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.23.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.30.0
sentence-transformers==3.4.1
setuptools==77.0.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
SQLAlchemy==2.0.39
streamlit==1.43.2
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.50.0
trio==0.29.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.3
zstandard==0.23.0
